arm64: dts: pm8150: Switch to software debouncing for pon keys
The hardware debouncing implementation in PM8150 is poorly implemented
and delays *all* key presses by the debounce delay, even when no
bouncing occurs. Let's switch to the kernel's software debouncing
implementation instead to reduce unnecessary kpdpwr and resin key press
delays, since it doesn't have this issue.

Note that we still have a 62 µs debounce delay added by the PMIC, since
it's not possible to fully disable hardware debouncing, but at least the
delay is much shorter than what it used to be.

This reduces the total wakeup latency (excluding deep sleep) from ~175
ms to ~160 ms, or about 9%, on a Pixel 2 XL with MSM8998 + PM8998.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>